Introduce Row Wrapper API.
This change allows to provide a row wrapper different from the default 
CursorRow. The main idea is to allow any consumer of the Database class 
to define how the query results should be turned into Python objects. A 
typical usage is that of extracting data from a database in the form of 
a Pandas DataFrame.